[build] workflow to pre-build config files as reference
Introduced a manually triggered workflow which will use a part of the code in PKGBUILD to generate the config files contained in the package later.
The workflow will force-push the result to a branch named after the source branch with the following pattern: 'configs/<source-branch-name>'.

It is used for documentation (and debugging) purposes only.