[btc-config] effective controller profiles will be combination results
1. New API function: applyGuideProfile
Reads the inbuilt 'GUIDE' profile and applies its contents to the
profile given as first argument. Output of result on stdout.
2. `.controls.lib`: calls applyGuideProfile when any controller profile
is selected and will create a temporary profile to use with AMX finally.

This process allows to maintain minimum functionality, prevents the user
from destroying these and makes the GUIDE functionality updateable as
well.