[configuration] added a way to add/remove simple values to arrays
The config import and data merge utility now recognizes 2 control
prefixes for special array manipulations.
These only work for arrays contained as substructures in dictionaries,
because they have to be attached to the key containing the array.

1. Prefix '@+': The entries of the update's array will be appended to
the base array of the same key name (minus the prefix). But they will
only be added if the base array does not already contain them. This
prefix can also create the array in base if there is no current value
associated with the key (values of other types won't be overriden).
2. Prefix '@-': Works like a filter. Any entry in the base array which
matches one of the entries in the update's array will be removed from
base.
3. It is possible to define both at once, since the effective key names
in the property files are different. They will be applied in iteration
order (which normally corresponds to the order of definition).

This array merge only fully works for simple data types because
json/javascript does not have a concept of object identities/classes, so
there's no way to match the base object against the corresponding update
object.
When nested objects are used, '@+' will be able to add them as new
entries instead of overwriting existing indices in base, but there will
be no recursive merge of object contents of 'identical' objects with
different indices in base and update.
'@-' will not work at all, because 2 simple objects will always be
different for js and there is no extended equals method based on any
kind of identity or similar.